+++ changelog.md
@@ -1,5 +1,10 @@
 Changelog
 
+### #5 2017-11-11 - Kent Huang
+- added fleet page update function
+- added ajax for fleet page update
+
+
 ### #4 2017-11-11 - Kabir Cahill
 - added fleet edit page
 - added fleet wackyid dropdown